requirements.txt
google-genai
mcp
requests
jupyter
ipykernel
langgraph
langchain_core
crewai[tools]
